Add pre_070 functions AND FIX RATING CALC FUNCTION ORDER!!
I confused the function for player overall with the score overall function. :facepalm: But honestly, the naming is as confusing as it can get. "AggregateSkillsets" and "AggregateRatings". What he hell. There are chart and player skillsets, just like there are chrat and player ratings; they're two names for the same thing. Argh